# openapi/babycharts-v1.yaml
openapi: 3.1.0
info:
  title: BabyCharts API
  description: Privacy-first self-hosted child growth, milestone and health tracking REST API.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Current API v1 endpoint
paths:
...
  /settings:
    get:
      summary: Get instance configuration settings
      security:
        - BearerAuth: []
        - SessionCookie: []
      responses:
        '200':
          description: Instance settings
    post:
      summary: Update instance configuration settings (admin only)
      security:
        - BearerAuth: []
        - SessionCookie: []
      responses:
        '200':
          description: Settings updated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SessionCookie:
      type: apiKey
      in: cookie
      name: babycharts_session